- added default to group.prio and group.parallelismDegree - added @Getter and @Setter to models - refactored filestructure to separate JPA models from other models
