+++ application-apis/groupapi/src/main/java/com/honeybadgers/groupapi/service/impl/GroupService.java
@@ -7,8 +7,8 @@ import com.honeybadgers.groupapi.service.IGroupService;
 import com.honeybadgers.models.exceptions.CreationException;
 import com.honeybadgers.models.exceptions.JpaException;
 import com.honeybadgers.models.exceptions.TransactionRetriesExceeded;
-import com.honeybadgers.models.model.Group;
-import com.honeybadgers.models.model.Task;
+import com.honeybadgers.models.model.jpa.Group;
+import com.honeybadgers.models.model.jpa.Task;
 import com.honeybadgers.models.exceptions.UnknownEnumException;
 import com.honeybadgers.postgre.repository.GroupRepository;
 import com.honeybadgers.postgre.repository.TaskRepository;
