+++ application-apis/groupapi/src/test/java/com/honeybadgers/groupapi/service/GroupConvertUtilsTest.java
@@ -3,8 +3,8 @@ package com.honeybadgers.groupapi.service;
 import com.honeybadgers.groupapi.models.GroupModel;
 import com.honeybadgers.groupapi.models.GroupModelActiveTimes;
 import com.honeybadgers.groupapi.service.impl.GroupConvertUtils;
-import com.honeybadgers.models.model.ActiveTimes;
-import com.honeybadgers.models.model.Group;
+import com.honeybadgers.models.model.jpa.ActiveTimes;
+import com.honeybadgers.models.model.jpa.Group;
 import com.honeybadgers.models.exceptions.UnknownEnumException;
 import com.honeybadgers.postgre.repository.GroupRepository;
 import org.junit.Before;
