+++ application-apis/groupapi/src/test/java/com/honeybadgers/groupapi/service/GroupServiceTest.java
@@ -4,14 +4,14 @@ import com.honeybadgers.communication.ICommunication;
 import com.honeybadgers.models.exceptions.CreationException;
 import com.honeybadgers.models.exceptions.JpaException;
 import com.honeybadgers.models.exceptions.TransactionRetriesExceeded;
-import com.honeybadgers.models.model.Task;
+import com.honeybadgers.models.model.jpa.Task;
 import com.honeybadgers.postgre.repository.GroupRepository;
 import com.honeybadgers.postgre.repository.TaskRepository;
 import org.hibernate.TransactionException;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import com.honeybadgers.models.model.Group;
+import com.honeybadgers.models.model.jpa.Group;
 import com.honeybadgers.groupapi.models.GroupModel;
 import com.honeybadgers.models.exceptions.UnknownEnumException;
 import com.honeybadgers.groupapi.service.impl.GroupService;
