+++ application-apis/taskapi/src/main/java/com/honeybadgers/taskapi/models/TaskModelActiveTimes.java
@@ -1,7 +1,7 @@
 package com.honeybadgers.taskapi.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.honeybadgers.models.model.ActiveTimes;
+import com.honeybadgers.models.model.jpa.ActiveTimes;
 import io.swagger.annotations.ApiModelProperty;
 
 import java.sql.Time;
