+++ application-apis/taskapi/src/main/java/com/honeybadgers/taskapi/service/ITaskConvertUtils.java
@@ -1,7 +1,7 @@
 package com.honeybadgers.taskapi.service;
 
-import com.honeybadgers.models.model.ActiveTimes;
-import com.honeybadgers.models.model.Task;
+import com.honeybadgers.models.model.jpa.ActiveTimes;
+import com.honeybadgers.models.model.jpa.Task;
 import com.honeybadgers.models.exceptions.UnknownEnumException;
 import com.honeybadgers.models.utils.IConvertUtils;
 import com.honeybadgers.models.exceptions.CreationException;
