+++ application-apis/taskapi/src/main/java/com/honeybadgers/taskapi/service/impl/TaskService.java
@@ -3,7 +3,8 @@ package com.honeybadgers.taskapi.service.impl;
 import com.honeybadgers.communication.ICommunication;
 import com.honeybadgers.communication.model.TaskQueueModel;
 import com.honeybadgers.models.exceptions.UnknownEnumException;
-import com.honeybadgers.models.model.*;
+import com.honeybadgers.models.model.jpa.Task;
+import com.honeybadgers.models.model.jpa.TaskStatusEnum;
 import com.honeybadgers.postgre.repository.GroupRepository;
 import com.honeybadgers.postgre.repository.TaskRepository;
 import com.honeybadgers.models.exceptions.CreationException;
