+++ application-cleaner/src/test/java/com/honeybadgers/cleaner/services/ScheduledServicesTest.java
@@ -1,10 +1,10 @@
 package com.honeybadgers.cleaner.services;
 
 import com.honeybadgers.communication.ICommunication;
-import com.honeybadgers.models.model.History;
-import com.honeybadgers.models.model.Paused;
-import com.honeybadgers.models.model.Task;
-import com.honeybadgers.models.model.TaskStatusEnum;
+import com.honeybadgers.models.model.jpa.History;
+import com.honeybadgers.models.model.jpa.Paused;
+import com.honeybadgers.models.model.jpa.Task;
+import com.honeybadgers.models.model.jpa.TaskStatusEnum;
 import com.honeybadgers.postgre.repository.PausedRepository;
 import com.honeybadgers.postgre.repository.TaskRepository;
 import org.junit.Test;
