+++ application-clienttests/src/main/java/com/honeybadgers/clienttests/models/TaskModelActiveTimes.java
@@ -1,7 +1,7 @@
 package com.honeybadgers.clienttests.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-//import com.honeybadgers.models.model.ActiveTimes;
+//import com.honeybadgers.models.model.jpa.ActiveTimes;
 import io.swagger.annotations.ApiModelProperty;
 
 import java.sql.Time;
