+++ application-clienttests/src/main/java/com/honeybadgers/clienttests/transactions/TransactionsService.java
@@ -3,8 +3,8 @@ package com.honeybadgers.clienttests.transactions;
 import com.honeybadgers.clienttests.models.ResponseModel;
 import com.honeybadgers.clienttests.models.TaskModel;
 import com.honeybadgers.communication.ICommunication;
-import com.honeybadgers.models.model.Task;
-import com.honeybadgers.models.model.TaskStatusEnum;
+import com.honeybadgers.models.model.jpa.Task;
+import com.honeybadgers.models.model.jpa.TaskStatusEnum;
 import com.honeybadgers.postgre.repository.TaskRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
