+++ application-models/src/main/java/com/honeybadgers/models/model/LockResponse.java
@@ -1,14 +1,14 @@
 package com.honeybadgers.models.model;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.sql.Timestamp;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
+@Setter
+@Getter
 public class LockResponse {
 
     private String name;
