+++ application-models/src/main/java/com/honeybadgers/models/model/jpa/ActiveTimes.java
@@ -1,14 +1,14 @@
-package com.honeybadgers.models.model;
+package com.honeybadgers.models.model.jpa;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.sql.Time;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
+@Setter
+@Getter
 public class ActiveTimes {
 
     private Time from;
