+++ application-models/src/main/java/com/honeybadgers/models/model/jpa/Group.java
@@ -1,4 +1,4 @@
-package com.honeybadgers.models.model;
+package com.honeybadgers.models.model.jpa;
 
 import com.vladmihalcea.hibernate.type.array.IntArrayType;
 import com.vladmihalcea.hibernate.type.json.JsonBinaryType;
@@ -16,10 +16,9 @@ import javax.validation.constraints.Min;
 import java.sql.Timestamp;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 
 @Entity
-@Table(name = "task")
+@Table(name = "\"group\"")
 @TypeDefs({
         @TypeDef(name = "jsonb", typeClass = JsonBinaryType.class),
         @TypeDef(name = "int-array", typeClass = IntArrayType.class)
@@ -28,20 +27,20 @@ import java.util.Map;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Task {
+public class Group {
 
     @Id
-    @Column(name = "id", unique = true, nullable = false)
+    @Column(name="id", unique = true, nullable = false)
     private String id;
 
     @ManyToOne
-    @JoinColumn(name = "group_id", nullable = false)
-    private Group group;
+    @JoinColumn(name = "parent_id")
+    private Group parentGroup;
 
     @Max(value = 9999)
     @Min(value = 0)
     @Column(name = "priority", nullable = false)
-    private int priority;
+    private int priority = 1;
 
     @Column(name = "deadline")
     private Timestamp deadline;
@@ -52,14 +51,9 @@ public class Task {
     private List<ActiveTimes> activeTimeFrames;
 
     // hibernate does not support boolean[] not even using hibernate-types-52
-    // -> boolean as int
     @Type(type = "int-array")
     @Column(name = "working_days", columnDefinition = "integer[]")
     private int[] workingDays;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "status", nullable = false)
-    private TaskStatusEnum status = TaskStatusEnum.Waiting;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "type_flag", nullable = false)
@@ -70,58 +64,38 @@ public class Task {
     private ModeEnum modeEnum = ModeEnum.Parallel;
 
     @Min(value = 0)
-    @Column(name = "retries")
-    private int retries = 0;
-
-    @Column(name = "force", nullable = false)
-    private boolean force = false;
+    @Column(name = "last_index_number")
+    private Integer lastIndexNumber;
 
     @Min(value = 1)
-    @Column(name = "index_number")
-    private Integer indexNumber;
+    @Column(name = "parallelism_degree")
+    private Integer parallelismDegree = 100;
 
-    @Type(type = "jsonb")
-    @Column(name = "meta_data", columnDefinition = "jsonb")
-    @Basic(fetch = FetchType.LAZY)
-    private Map<String, String> metaData;
-
-    @Column(name = "total_priority")
-    private Long totalPriority;
-
-    @Type(type = "jsonb")
-    @Column(name = "history", columnDefinition = "jsonb")
-    @Basic(fetch = FetchType.LAZY)
-    private List<History> history;
+    @Column(name = "current_parallelism_degree", nullable = false)
+    private Integer currentParallelismDegree = 0;
 
 
     @PrePersist
     void checkModeParameters() {
-        if(this.modeEnum == ModeEnum.Sequential) {
-            assert this.indexNumber != null;
-        } else if(this.modeEnum == ModeEnum.Parallel) {
-            // TODO decide if necc
-            assert this.group.getParallelismDegree() != null;
+        if(this.modeEnum == ModeEnum.Sequential && this.lastIndexNumber == null) {
+            this.lastIndexNumber = 0;
         }
     }
 
     @Override
     public String toString() {
-        return "Task{" +
+        return "Group{" +
                 "id='" + id + '\'' +
-                ", group=" + group +
+                ", parentGroup=" + parentGroup +
                 ", priority=" + priority +
                 ", deadline=" + deadline +
                 ", activeTimeFrames=" + activeTimeFrames +
                 ", workingDays=" + Arrays.toString(workingDays) +
-                ", status=" + status +
                 ", typeFlagEnum=" + typeFlagEnum +
                 ", modeEnum=" + modeEnum +
-                ", retries=" + retries +
-                ", force=" + force +
-                ", indexNumber=" + indexNumber +
-                ", metaData=" + metaData +
-                ", totalPriority=" + totalPriority +
-                ", history=" + history +
+                ", lastIndexNumber=" + lastIndexNumber +
+                ", parallelismDegree=" + parallelismDegree +
+                ", currentParallelismDegree=" + currentParallelismDegree +
                 '}';
     }
 }
