+++ application-models/src/main/java/com/honeybadgers/models/model/jpa/History.java
@@ -1,14 +1,14 @@
-package com.honeybadgers.models.model;
+package com.honeybadgers.models.model.jpa;
 
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.sql.Timestamp;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
+@Setter
+@Getter
 public class History {
 
     private String status;
