+++ application-models/src/main/java/com/honeybadgers/models/model/jpa/Paused.java
@@ -1,4 +1,4 @@
-package com.honeybadgers.models.model;
+package com.honeybadgers.models.model.jpa;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
