+++ application-postgrecommunication/src/main/java/com/honeybadgers/postgre/repository/GroupRepository.java
@@ -1,7 +1,7 @@
 package com.honeybadgers.postgre.repository;
 
 
-import com.honeybadgers.models.model.Group;
+import com.honeybadgers.models.model.jpa.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
