+++ application-postgrecommunication/src/main/java/com/honeybadgers/postgre/repository/TaskRepository.java
@@ -1,7 +1,7 @@
 package com.honeybadgers.postgre.repository;
 
 
-import com.honeybadgers.models.model.Task;
+import com.honeybadgers.models.model.jpa.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
